Don't allow outside forces to modify argc or argv
Google Test likes to completely overwrite anything in argv that has
a Google Test argument, but we don't want anything to be able to be
able to directly overwrite argv. Other libraries may need or want an
unmodified version. This change allows anyone to grab a copy of argv
and modify that to their heart's content :^)