Fix and improve an issue with string formatting
For some reason, l-value references were being uplifted to r-value
references without the use of MoveTemp or std::move. This was causing
types to be unintentionally moved when passed as non-cost references for
string formatting

This change also makes the value being formatted const. There's no
valid reason for string formatters to need to modify the target value

Unfortunately this change also adds a const-specific Forward overload,
which it doesn't seem like the standard library has to do